services/ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.0.3
kafka-python==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
boto3==1.35.98
alembic==1.13.1
structlog==23.2.0